Improve outputs. Add comments. Dedup code. Move pkg/kubernetes into pkg/manifest to dedup dependencies.
