+++ cmd/mesh/shared.go
@@ -21,6 +21,7 @@ import (
 	"io"
 	"os"
 
+	"istio.io/operator/pkg/manifest"
 	"istio.io/pkg/log"
 )
 
@@ -99,4 +100,13 @@ func (l *logger) print(s string) {
 
 func refreshGoldenFiles() bool {
 	return os.Getenv("REFRESH_GOLDEN") == "true"
+}
+
+// getKubeExecClient returns a Kubernetes client specified by kubeConfig and configContext
+func getKubeExecClient(kubeConfig, configContext string) manifest.ExecClient {
+	kubeClient, err := manifest.NewClient(kubeConfig, configContext)
+	if err != nil {
+		l.logAndFatalf("Abort. Failed to connect Kubernetes API server: %v", err)
+	}
+	return kubeClient
 }
